w changes
