+++ src/main/java/frc/robot/subsystems/pivot/PivotIOTalonFX.java
@@ -132,22 +132,12 @@ public class PivotIOTalonFX implements PivotIO {
     inputs.followerSupplyCurrentAmps = followerPivotMotor.getSupplyCurrent().getValueAsDouble();
   }
 
-  /**
-   * Sets the voltage of the pivot motors
-   *
-   * @param volts the voltage
-   */
   @Override
   public void setVoltage(double volts) {
     leaderPivotMotor.setControl(voltageControl.withOutput(volts));
     followerPivotMotor.setControl(voltageControl.withOutput(volts));
   }
 
-  /**
-   * Gets the angle of the pivot
-   *
-   * @return angle of pivot in rotations
-   */
   @Override
   public double getAngle() {
     pivotPos.refresh();
@@ -159,11 +149,6 @@ public class PivotIOTalonFX implements PivotIO {
     return Math.abs(pivotTargetAngle - getAngle()) < PivotConstants.PIVOT_ACCEPTABLE_ERROR;
   }
 
-  /**
-   * Sets the output of the pivot
-   *
-   * @param output output value from -1.0 to 1.0
-   */
   @Override
   public void setPivotSpeed(double output) {
     leaderPivotMotor.set(output);
